Add support for horizontal stack views.
This change adds the ability to change the `axis` of an
AloeStackView between vertical and horizontal directions. The
`axis` can be changed dynamically, and separators become vertical
when the `axis` is `.horizontal`.

This change combines and builds on the work of PRs #15 and #23, and
addresses issue #13.

Special thanks to @neobeppe and @apptekstudios for their proposals
for implementing this feature.