Change StackViewCell to pass touches to UIControls correctly.
This change fixes an additional edge case with PR #55. In the previous
implementation, StackViewCell would only pass touches to UIControls if the touch
occurred on the UIControl itself, not on any of its subviews.

This meant that UIControls such as UISwitch would not receive touches if the row
had a tap handler set on it.

This change addresses the issue by searching up the superview chain for any
parent UIControls before allowing the touches to go to the tap handler's gesture
recognizer.